+++ frontend/src/Login.css
@@ -1,0 +1,70 @@
+:root {
+  font-family: "Karla", sans-serif;
+  font-size: 16px;
+}
+
+.row {
+  display: flex;
+}
+
+.column {
+  flex: 50%;
+  height: 100vh;
+}
+
+.dark-purple {
+  background-color: #330c7a;
+}
+
+h1 {
+  padding-left: 1em;
+  padding-top: 0.5em;
+  font-size: 3em;
+  font-family: "Open Sans", sans-serif;
+  color: #fff;
+}
+
+a {
+  text-decoration: none;
+  color: #7140cc;
+}
+
+h2 {
+  font-family: "Montserrat", sans-serif;
+  font-size: 36px;
+  text-align: center;
+}
+
+.login {
+  margin: auto;
+  width: 50%;
+  margin-top: 150px;
+}
+
+input[type="submit"] {
+  font-family: "Karla";
+  display: block;
+  margin: 0 auto;
+  background-color: #330c7a;
+  color: #fff;
+  border-radius: 15px;
+  font-size: 2em;
+  width: 100%;
+  margin-top: 20px;
+}
+
+h3 {
+  text-align: left;
+}
+
+input[type="text"] {
+  background-color: #d9d9d9;
+  text-align: left;
+  width: 100%;
+  outline: none;
+  border-radius: 15px;
+}
+
+p {
+  text-align: center;
+}
